Move save method logic from ApiController into new autosave-controller to match route changes
